+++ lib/helpers/base_helper.rb
@@ -50,8 +50,4 @@ module BaseHelper
   def vk_like_btn
     '<div id="vk_like" class="pull-left"></div>'
   end
-
-  def vk_comments_box(uid)
-    "<div class=\"VKborders\"><div id=\"vk_comments\" data-uid=\"#{uid}\"></div></div>"
-  end
 end